fix(react-account): align account self-service MSW handlers with backend routes
createAccountHandlers (react-account/testing) had drifted from the backend
contract and the @granit/account API on 9 self-service endpoints — wrong
routes, methods and status codes:

  POST /password/change      -> POST /change-password (204)
  POST /password/forgot      -> POST /forgot-password (202)
  POST /password/reset       -> POST /reset-password (204)
  POST /register (201, body) -> POST /register (202, no body, 409 on conflict)
  POST /confirm-email        -> GET /confirm-email (204)
  POST /resend-confirmation  -> POST /resend-confirmation-email (202)
  POST /email/change         -> POST /change-email (202)
  POST /email/confirm-change -> POST /confirm-email-change (204)
  DELETE /account            -> POST /delete (202)

The mocks never matched the real calls; forgot/reset password only worked
in the showcase because createLocalAuthHandlers carried vestigial account
"orphan" handlers (with the correct routes) that compensated for the drift.

Trim createLocalAuthHandlers to login-only (credential login, two-factor
login, passkey assertion). Account self-service flows belong exclusively to
createAccountHandlers. Status codes now mirror the Identity.Local OpenAPI.

Verified: tsc + ESLint clean, 75 react-account/react-authentication-local
tests pass, 28 showcase auth-local tests pass.
